requirements.txt
keras==2.2.4
tensorflow-gpu==1.9.0
deepbrain==0.1
scikit-image==0.15.dev0
nibabel==2.3.2
numpy==1.15.4
matplotlib==3.0.2
opencv-contrib-python==3.4.2.16
